Experimenting with tabnine
I've been experimenting with tabnine for a bit, and it works
alright to automatically complete text (like comments or commit messages)

Code completion is hit or miss. And it is pretty annoying that it just
compltes text instead of working with the language server.
That way autoimports and things like that would work as well.
Instead it just compltes 'text', so no docs, no autoimport, and so on.

Also it drains quite a bit of battery, so I might switch it off on the
go.