feat(tmux): Add bash script that automatically names new tmux session
Add a bash script that automatically names the new tmux session the same
as the directory I'm in.
If there already is a tmux session with that name, a number is appended
to make it unique.
It also removes ':' because tmux can't handle those.

I'm not sure if I'll keep it that way, since I generally don't need more
tmux sessions in the same directory but I thought I'd try it out.
Maybe I'll revert back to my old alias though (that just names the
session, and tmux will throw an error if there already is a session
with that name).