feat: Add custom session names
When invoking the script, a name can be passed as the first argument,
and the session will be named using the string.
If there are no arguments, it works as before, and the name of the
current folder will be used as the session name.

NOTE: custom names are currently not checked in the same way directory
names are. So in case a session by that name already exists, the script
will crash.