feat: Disable TabNine
Disabling TabNine to see whether it's really needed.
If it's not, I will permanently remove it since it consumes a decent
amount of resources on this device.
This plugin also isn't actively developed anymore, so I don't know
whether there are still updates. TabNine doesn't seem to have gotten any
better anyway over the last year or two.
There is also an official
[plugin](https://github.com/codota/tabnine-nvim) by the TabNine people
but that one sucks, since it doesn't integrate with cmp, instead
competing with it (for keyboard shortcuts for example).