feat: Remove codi
I don't use it, ever, and I don't even know how it works/it doesn't seem
like it works.
It also has known bugs in the FAQ, but there has been no commit in 2
years, so I assume the project is abandoned while buggy, hence there is
no reason to keep it around (it's not like I'll be missing anything
anyway).